resolve mahmod's code

+++ src/main/java/com/example/SearchEngine/invertedIndex/TrieInvertedIndex.java
@@ -45,7 +45,7 @@ public class TrieInvertedIndex implements InvertedIndex {
     }
 
 
-    private void indexer(TrieNode root, Integer documentId, String fieldName, List<Token> tokens) {
+    private void indexer(TrieNode root, Long documentId, String fieldName, List<Token> tokens) {
         for (Token token : tokens) {
             TrieNode lastNode = getWordLastNode(root, token);
             lastNode.setEndOfTerm();
@@ -53,7 +53,7 @@ public class TrieInvertedIndex implements InvertedIndex {
         }
     }
 
-    public void remover(List<Token> tokens, TrieNode root, Integer documentId) {
+    public void remover(List<Token> tokens, TrieNode root, Long documentId) {
         for (Token token : tokens) {
             if (checkWordExist(root, token)) {
                 TrieNode lastNode = getWordLastNode(root, token);
@@ -65,29 +65,29 @@ public class TrieInvertedIndex implements InvertedIndex {
     @Override
     public void addDocument(String schemaName, Map<String, Object> document) throws Exception {
         TrieNode root = SchemaRoot.getSchemaRoot(schemaName);
-        CollectionInfo.insertDocument(schemaName, (Integer) document.get("id"));
+        CollectionInfo.insertDocument(schemaName, (Long) document.get("id"));
         for (String fieldName : document.keySet()) {
             List<Token> tokens = getTokens(schemaName, fieldName, document);
             if (!tokens.isEmpty()) {
                 fuzzyTrie.addField((String) document.get(fieldName), schemaName);
-                CollectionInfo.addField(schemaName, (Integer) document.get("id"), tokens);
+                CollectionInfo.addField(schemaName, (Long) document.get("id"), tokens);
             }
-            indexer(root, (Integer) document.get("id"), fieldName, tokens);
+            indexer(root, (Long) document.get("id"), fieldName, tokens);
         }
     }
 
     @Override
     public void deleteDocument(String schemaName, Map<String, Object> document) throws Exception {
-        CollectionInfo.removeDocument(schemaName, (Integer) document.get("id"));
+        CollectionInfo.removeDocument(schemaName, (Long) document.get("id"));
         TrieNode root = SchemaRoot.getSchemaRoot(schemaName);
 
         for (String fieldName : document.keySet()) {
             List<Token> tokens = getTokens(schemaName, fieldName, document);
             if (!tokens.isEmpty()) {
                 fuzzyTrie.removeField((String) document.get(fieldName), schemaName);
-                CollectionInfo.removeField(schemaName, (Integer) document.get("id"), tokens);
+                CollectionInfo.removeField(schemaName, (Long) document.get("id"), tokens);
             }
-            remover(tokens, root, (Integer) document.get("id"));
+            remover(tokens, root, (Long) document.get("id"));
         }
     }
