+++ src/main/java/com/example/SearchEngine/invertedIndex/utility/CollectionInfo.java
@@ -11,12 +11,12 @@ import java.util.List;
 import java.util.Map;
 
 public class CollectionInfo {
-    private static Map<String, Integer> numberOfDocument = new HashMap<>();
+    private static Map<String, Long> numberOfDocument = new HashMap<>();
     private static Map<String, Double> documentsTotalLength = new HashMap<>();
-    private static Map<String, HashSet<Integer>> allDocuments = new HashMap<>();
-    private static Map<String, Map<Integer, Double>> documentLength = new HashMap<>();
+    private static Map<String, HashSet<Long>> allDocuments = new HashMap<>();
+    private static Map<String, Map<Long, Double>> documentLength = new HashMap<>();
 
-    public static Double getDocumentLength(String schemaName, Integer documentId) {
+    public static Double getDocumentLength(String schemaName, Long documentId) {
         return documentLength.getOrDefault(schemaName, new HashMap<>())
                 .getOrDefault(documentId, 0.0);
     }
@@ -25,18 +25,18 @@ public class CollectionInfo {
         return documentsTotalLength.getOrDefault(schemaName, 1.0);
     }
 
-    public static void insertDocument(String schemaName, Integer documentId) {
+    public static void insertDocument(String schemaName, Long documentId) {
         allDocuments.putIfAbsent(schemaName, new HashSet<>());
         allDocuments.get(schemaName).add(documentId);
         updateNumberOfDocument(schemaName, 1);
     }
 
-    public static boolean isDocumentExist(String schemaName, Integer documentId) {
+    public static boolean isDocumentExist(String schemaName, Long documentId) {
         allDocuments.putIfAbsent(schemaName, new HashSet<>());
         return allDocuments.get(schemaName).contains(documentId);
     }
 
-    public static void removeDocument(String schemaName, Integer documentId) {
+    public static void removeDocument(String schemaName, Long documentId) {
         if (allDocuments.get(schemaName).contains(documentId)) {
             allDocuments.get(schemaName).remove(documentId);
             updateNumberOfDocument(schemaName, -1);
@@ -44,7 +44,7 @@ public class CollectionInfo {
     }
 
 
-    public static void updateDocumentLength(String schemaName, Integer documentId, Double length) {
+    public static void updateDocumentLength(String schemaName, Long documentId, Double length) {
         documentLength.putIfAbsent(schemaName, new HashMap<>());
         documentLength.get(schemaName).put(documentId, documentLength.get(schemaName).getOrDefault(documentId, 0.0) + length);
         documentsTotalLength.put(schemaName, documentsTotalLength.getOrDefault(schemaName, 0.0) + length);
@@ -52,21 +52,21 @@ public class CollectionInfo {
 
 
     public static void updateNumberOfDocument(String schemaName, Integer value) {
-        numberOfDocument.putIfAbsent(schemaName, 0);
+        numberOfDocument.putIfAbsent(schemaName, 0L);
         numberOfDocument.put(schemaName, numberOfDocument.get(schemaName) + value);
     }
 
-    public static Integer getNumberOfDocument(String schemaName) {
-        return numberOfDocument.getOrDefault(schemaName, 0);
+    public static Long getNumberOfDocument(String schemaName) {
+        return numberOfDocument.getOrDefault(schemaName, 0L);
     }
 
-    public static void addField(String schemaName, Integer documentId, List<Token> tokens) {
+    public static void addField(String schemaName, Long documentId, List<Token> tokens) {
         Integer length = tokens.size();
         Double weight = tokens.get(0).getWeight();
         updateDocumentLength(schemaName, documentId, weight * length);
     }
 
-    public static void removeField(String schemaName, Integer documentId, List<Token> tokens) {
+    public static void removeField(String schemaName, Long documentId, List<Token> tokens) {
         Integer length = tokens.size();
         Double weight = tokens.get(0).getWeight();
         updateDocumentLength(schemaName, documentId, -1 * weight * length);
@@ -102,9 +102,9 @@ public class CollectionInfo {
     }
 
     public static void load() throws Exception {
-        numberOfDocument = (Map<String, Integer>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "numberOfDocument");
+        numberOfDocument = (Map<String, Long>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "numberOfDocument");
         documentsTotalLength = (Map<String, Double>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "documentsTotalLength");
-        documentLength = (Map<String, Map<Integer, Double>>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "documentLength");
-        allDocuments = (Map<String, HashSet<Integer>>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "allDocuments");
+        documentLength = (Map<String, Map<Long, Double>>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "documentLength");
+        allDocuments = (Map<String, HashSet<Long>>) loadAttribute(Constants.Paths.SCHEMA_STORAGE_PATH + "allDocuments");
     }
 }
