+++ src/main/java/com/example/SearchEngine/schema/log/TrieLogLoader.java
@@ -27,13 +27,13 @@ public class TrieLogLoader {
         String line;
         while((line = reader.readLine()) != null) {
             String[] words = line.split(" ");
-            Map<String, Object> document = documentStorageService.getDocument(schemaName, Integer.parseInt(words[2]));
+            Map<String, Object> document = documentStorageService.getDocument(schemaName, Long.parseLong(words[2]));
             if (Command.valueOf(words[1]) == Command.INSERT){
                 trieInvertedIndex.addDocument(schemaName, document);
                 documentFilterService.addDocument(schemaName, (HashMap<String, Object>) document);
             } else if (Command.valueOf(words[1]) == Command.DELETE){
                 trieInvertedIndex.deleteDocument(schemaName, document);
-                documentFilterService.removeDocument(schemaName, (HashMap<String, Object>) documentStorageService.getDocument(schemaName, Integer.parseInt(words[2])));
+                documentFilterService.removeDocument(schemaName, (HashMap<String, Object>) documentStorageService.getDocument(schemaName, Long.parseLong(words[2])));
             } else if (Command.valueOf(words[1]) == Command.UPDATE){
                 // Update method from DocumentStorageService
             }
